better library grid view
