+++ lib/features/library/library_items.dart
@@ -1,19 +1,13 @@
-import 'package:abs_api/abs_api.dart';
-import 'package:abs_flutter/generated/l10n.dart';
+import 'package:abs_flutter/features/library/items/library_item_widget.dart';
 import 'package:abs_flutter/models/library_preview.dart';
-import 'package:abs_flutter/models/library_preview_item.dart';
 import 'package:abs_flutter/provider/connection_provider.dart';
 import 'package:abs_flutter/provider/library_items_provider.dart';
 import 'package:abs_flutter/provider/progress_provider.dart';
-import 'package:abs_flutter/provider/user_provider.dart';
-import 'package:abs_flutter/widgets/album_image.dart';
+import 'package:abs_flutter/widgets/error_text.dart';
 import 'package:abs_flutter/widgets/no_connection.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:go_router/go_router.dart';
-import 'package:shimmer/shimmer.dart';
-import 'package:abs_flutter/models/user.dart' as m;
+import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
 
 class LibraryItems extends ConsumerStatefulWidget {
   const LibraryItems({super.key});
@@ -68,7 +62,7 @@ class _LibraryItemsState extends ConsumerState<LibraryItems> {
     final libraryItems = ref.watch(libraryItemsProvider);
 
     if (!connection) {
-      return _buildSafeArea(const NoConnection());
+      return const NoConnection();
     }
 
     (MediaQuery.of(context).size.width < 900)
@@ -82,201 +76,32 @@ class _LibraryItemsState extends ConsumerState<LibraryItems> {
     }
 
     if (libraryItems == null) {
-      return _buildSafeArea(_buildShimmerLoading());
+      return const ErrorText('No library items found');
     } else {
       if (libraryItems.page <= 0) {
         final progressProv = ref.read(progressProvider);
         progressProv.getAllProgress();
       }
       _hasMore = libraryItems.total != libraryItems.items.length;
-      return _buildSafeArea(_buildItems(context, libraryItems));
+      return _buildItems(context, libraryItems);
     }
   }
 
-  Widget _buildSafeArea(Widget child) {
-    return SafeArea(child: child);
-  }
-
-  Widget _buildItems(BuildContext context, LibraryPreview items) {
-    if (items.items.isEmpty) {
-      return Center(child: PlatformText(S.of(context).noItemsFound));
-    }
-
+  Widget _buildItems(BuildContext context, LibraryPreview libraryItems) {
     return LayoutBuilder(
-      builder: (context, constraints) {
-        final crossAxisCount = (constraints.maxWidth / 200).floor();
-        return GridView.builder(
+      builder: (BuildContext context, BoxConstraints constraints) {
+        return AlignedGridView.count(
+          crossAxisCount: constraints.maxWidth ~/ 175,
+          mainAxisSpacing: 12,
+          crossAxisSpacing: 8,
+          padding: EdgeInsets.only(top: abovePadding, left: 8, right: 8),
+          itemCount: libraryItems.items.length,
           controller: _scrollController,
-          padding: EdgeInsets.only(
-              left: 8.0, right: 8.0, top: abovePadding, bottom: 8.0),
-          gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-            crossAxisCount: crossAxisCount,
-            crossAxisSpacing: 8.0,
-            mainAxisSpacing: 2.0,
-            childAspectRatio: 0.75,
-          ),
-          itemCount: items.items.length + (_isLoadingMore ? 20 : 0),
           itemBuilder: (context, index) {
-            if (index >= items.items.length) {
-              return _buildShimmerPlaceholder();
-            }
-            final item = items.items[index];
-            return _buildItemCard(context, item);
+            return LibraryItemWidget(item: libraryItems.items[index]);
           },
         );
       },
-    );
-  }
-
-  Widget _buildItemCard(BuildContext context, LibraryPreviewItem item) {
-    final currentUser = ref.watch(currentUserProvider);
-
-    return MouseRegion(
-      cursor: SystemMouseCursors.click,
-      child: GestureDetector(
-        onTap: () {
-          context.push('/view/book/${item.id}');
-        },
-        child: Padding(
-          padding: const EdgeInsets.only(bottom: 8.0),
-          child: PlatformWidget(
-            material: (_, __) => InkWell(
-              onTap: () {
-                context.push('/view/${item.mediaType}/${item.id}');
-              },
-              child: _buildCard(item, currentUser),
-            ),
-            cupertino: (_, __) => GestureDetector(
-              onTap: () {
-                context.push('/view/${item.mediaType}/${item.id}');
-              },
-              child: _buildCard(item, currentUser),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildCard(LibraryPreviewItem item, m.User? currentUser) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        AspectRatio(
-          aspectRatio: 1.0,
-          child: ClipRRect(
-            borderRadius: BorderRadius.circular(8.0),
-            child: Semantics(
-              label: S.of(context).bookCover,
-              child: Stack(
-                children: [
-                  AspectRatio(
-                    aspectRatio: 1.0,
-                    child: AlbumImage(item.id),
-                  ),
-                  if (item.mediaType == 'book')
-                    Consumer(builder: (context, ref, child) {
-                      MediaProgress? progress = ref.watch(
-                          progressProviderWithItemId(
-                              ItemEpisodeId(item.id, null)));
-                      return Column(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          item.seriesLabel != null
-                              ? Align(
-                                  alignment: Alignment.topRight,
-                                  child: Padding(
-                                      padding: const EdgeInsets.all(4.0),
-                                      child: Chip(
-                                          visualDensity: const VisualDensity(
-                                              horizontal:
-                                                  VisualDensity.minimumDensity,
-                                              vertical:
-                                                  VisualDensity.minimumDensity),
-                                          label: PlatformText(
-                                              '#${item.seriesLabel!}'))),
-                                )
-                              : const SizedBox.shrink(),
-                          LinearProgressIndicator(
-                            value: progress?.progress?.toDouble() ?? 0,
-                            minHeight: 5.0,
-                            valueColor: const AlwaysStoppedAnimation<Color>(
-                                Colors.green),
-                            backgroundColor: Colors.grey[300],
-                          ),
-                        ],
-                      );
-                    }),
-                ],
-              ),
-            ),
-          ),
-        ),
-        const SizedBox(height: 8.0),
-        Text(
-          item.title,
-          textAlign: TextAlign.center,
-          style: const TextStyle(fontWeight: FontWeight.bold),
-          overflow: TextOverflow.ellipsis,
-          maxLines: 1,
-        ),
-        Text(
-          item.authors.join(", "),
-          textAlign: TextAlign.center,
-          overflow: TextOverflow.ellipsis,
-          maxLines: 1,
-        ),
-      ],
-    );
-  }
-
-  Widget _buildShimmerLoading() {
-    return LayoutBuilder(
-      builder: (context, constraints) {
-        final crossAxisCount = (constraints.maxWidth / 200).floor();
-        return GridView.builder(
-          padding: EdgeInsets.only(
-              left: 8.0, right: 8.0, top: abovePadding, bottom: 8.0),
-          gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-            crossAxisCount: crossAxisCount,
-            crossAxisSpacing: 8.0,
-            mainAxisSpacing: 8.0,
-            childAspectRatio: 0.7,
-          ),
-          itemCount: 20,
-          itemBuilder: (context, index) => _buildShimmerPlaceholder(),
-        );
-      },
-    );
-  }
-
-  Widget _buildShimmerPlaceholder({bool withText = true}) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        AspectRatio(
-          aspectRatio: 1.0,
-          child: Shimmer.fromColors(
-            baseColor: Colors.grey[300]!,
-            highlightColor: Colors.grey[100]!,
-            child: Container(color: Colors.white),
-          ),
-        ),
-        if (withText) ...[
-          const SizedBox(height: 8.0),
-          Shimmer.fromColors(
-            baseColor: Colors.grey[300]!,
-            highlightColor: Colors.grey[100]!,
-            child: Container(height: 20.0, color: Colors.white),
-          ),
-          const SizedBox(height: 4.0),
-          Shimmer.fromColors(
-            baseColor: Colors.grey[300]!,
-            highlightColor: Colors.grey[100]!,
-            child: Container(height: 20.0, width: 50, color: Colors.white),
-          ),
-        ],
-      ],
     );
   }
